Extract imagemaid analyzer + progress snapshot into focused modules
Sprint 4, PR #1.  Applies the Sprint 3 logscan.py playbook to
logscan_imagemaid_analysis.py.  Two of its twelve functions
accounted for 697 of the 1020 lines; both are now in their
own modules.

Public API in modules/logscan_imagemaid_analyzer:
- analyze_imagemaid_log_content(content, log_path=None)
    468-line stateful parser that walks an imagemaid run log
    line-by-line, tracking Quickstart run markers, runtime
    timings, per-operation state, and error/critical log-level
    counts.  Returns {'summary': ..., 'recommendations': ...}.

Public API in modules/logscan_imagemaid_progress:
- build_imagemaid_progress_snapshot(summary=None)
    229-line pure function that turns a parsed imagemaid summary
    into a UI-friendly progress payload (table rows for each
    operation with per-row status badges + runtime cells).

This is a PURE MOVE.  No parser logic changed, no branches
reordered.  Every summary field and recommendation is
byte-identical.

Circular-import avoidance: Rather than have
logscan_imagemaid_analysis re-export the two extracted functions
(which would create an import cycle since the analyzer needs the
helpers that live in analysis), quickstart.py's import of
analyze_imagemaid_log_content moves to the new module directly.
build_imagemaid_progress_snapshot has no external callers -- only
the analyzer uses it -- so no import updates needed there.

Ruff auto-cleaned the now-unused imports in
logscan_imagemaid_analysis.py (hashlib, all three logscan_resume
formatters that moved to the progress module).

Impact:
- modules/logscan_imagemaid_analysis.py: 1020 -> 316 (-704 / -69.0%)
- modules/logscan_imagemaid_analyzer.py: NEW 510 lines
- modules/logscan_imagemaid_progress.py: NEW 256 lines
- quickstart.py: +2 lines (split one import into two)

All 1284 tests pass.  Ruff + black clean.
